Support subprocess_test_runner on Windows
This allows tests to be run in a child process. Since Windows doesn't allow
forking, we simply execute the test binary again, passing in special arguments
to tell it to run a single test and report messages on a special pipe. We also
add a Windows version of mettle's POSIX-based read_into, which uses async I/O
to simultaneously read from multiple pipes (until we receive an event telling
us to stop - currently just the process dying).