Simplify the logic for the make_timeout_monitor function
make_timeout_monitor no longer requires you to create a new process group
before calling it; instead, you create the process group afterwards and use a
pipe to tell the parent what the pgid is. This removes a lot of convoluted
code that wasn't really necessary.

This commit also features more-thorough tests to verify that we're killing
process groups correctly.