Be paranoid about receiving the pgid from the test process
This resolves a deadlock that would (hopefully) never have cropped up: if the
test process failed to send its pgid to the parent process, the recv_pgid
call would have blocked indefinitely because we were holding a write fd open.
Now, we close the write fd earlier so this should be resolved (but the
recv_pgid would still hang for as long as the test's timeout, since the
monitor and timer processes have their own write fds).